+++ back/src/routes/products.js
@@ -4,67 +4,6 @@ const axios = require('axios');
 const { allProducts, dbData, todaInfo } = require('../controllers');
 const { Product } = require('../db');
 
-// const dataBaseTesting = [
-
-//   (item1 = {
-//     name: 'nombrePrueba1',
-//     category: 'categoriaPrueba1',
-//     price: 1,
-//     stock: 1,
-//     published: 1,
-//   }),
-//   (item2 = {
-//     name: 'nombrePrueba2',
-//     category: 'categoriaPrueba2',
-//     price: 1,
-//     stock: 1,
-//     published: 1,
-//   }),
-//   (item3 = {
-//     name: 'nombrePrueba3',
-//     category: 'categoriaPrueba3',
-//     price: 1,
-//     stock: 1,
-//     published: 1,
-//   }),
-//   (item4 = {
-//     name: 'nombrePrueba4',
-//     category: 'categoriaPrueba4',
-//     price: 1,
-//     stock: 1,
-//     published: 1,
-//   }),
-//   (item5 = {
-//     name: 'nombrePrueba5',
-//     category: 'categoriaPrueba5',
-//     price: 1,
-//     stock: 1,
-//     published: 1,
-//   }),
-//   (item6 = {
-//     name: 'nombrePrueba6',
-//     category: 'categoriaPrueba6',
-//     price: 1,
-//     stock: 1,
-//     published: 1,
-//   }),
-// ];
-
-// dataBaseTesting.map(
-//   async (el) =>
-//     await Product.create({
-//       name: el.name,
-//       category: el.category,
-//       price: el.price,
-//       stock: el.stock,
-//       published: el.published,
-//     })
-// );
-// const mostrar = async () => {
-//   console.log(await Product.findAll());
-// };
-// mostrar();
-// || /productos || //
 router.get('', async (req, res) => {
   try {
     const products = await todaInfo();
@@ -74,7 +13,6 @@ router.get('', async (req, res) => {
   }
 });
 
-// || /productos/id || //
 router.get('/:id', async (req, res) => {
   const id = req.params.id;
   const products = await todaInfo();
@@ -112,12 +50,17 @@ router.post('', async (req, res) => {
 router.delete('/:id', async (req, res) => {
   try {
     const { id } = req.params;
-    await Product.destroy({
-      where: {
-        id,
-      },
-    });
-    res.sendStatus(204);
+    const product = await Product.findByPk(id);
+    if (product !== null) {
+      await Product.destroy({
+        where: {
+          id,
+        },
+      });
+      res.sendStatus(204);
+    } else {
+      throw new Error('the id does not exist');
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
